initial commit
I haven't made any of the site, not even a README.md, but I want to be able to test things on the Nintendo 3DS Internet Browser using the /x/eval.html file right away! I want to have easter eggs in the DS Day section that may be accessed on the Nintendo 3DS, you see. ;)