try to fix /x/eval console
The new /x/eval code did work on my 3DS, except the console, which would crash. I suspect that's because it doesn't support setting the style property of an element, so I've replaced that bit of code. I also adjusted the style of console.error and removed the border on a fatal error message, since it's basically like a return value except there is no value, only disappointment. I also added a README.md file.