more /x/eval tweaks
I think I've heard it's bad to have such frequent commits to change the same exact thing, but I have to make a commit to test my changes. I think this may be the last update to /x/eval for now. The console is now passed as an argument to your code, leaving the normal console untouched. I also tweaked the style of console.info and added a .gitignore so I can take notes on what I should put in these excessively long commit messages.