more ds day work
I changed the way the DS Day logo header works because the old way didn't work on 3DS, removed dark mode from the DS Day site because black-on-white is more on-brand, styled the navigation, and wrote some text for the About page. I also updated /x/eval AGAIN with the change mentioned previously that the body expands to fit the console and stays that way rather than always being super long and changed the default code to something that better demonstrate's the page's features. Now, if you'll excuse me, I am going to test these changes on 3DS.