simpler approach
It won't look exactly how it's supposed to on 3DS if it works, but at least it won't look absolutely terrible. I reverted the viewport meta tag, changed how the navigation style works, and visually tweaked it a bit now that I can. Before, the viewport was very wide and the navigation was the widest thing in it, so I think that may have been the problem. The logo was also stretched sometimes, but I'll get to that later.