resize /pics/ds-day.svg
The navigation is now correct, but the banner is still sometimes the wrong size. It appears that the browser is not respecting the width and height attributes when the page first loads... or something? Anyway, I resized it with an online SVG editor to exactly 300 pixels wide, which is significantly less than the 3DS's 320px bottom screen width. Hopefully this fixes it!